move blocks.go into environment subpackage
also renamed some variables for better searchability.
