+++ fvm/environment/blocks_test.go
@@ -1,4 +1,4 @@
-package fvm
+package environment
 
 import (
 	"testing"
